requirements.txt
openai>=0.27.0
streamlit>=1.25.0
python-dotenv>=0.21.0